refactor: 패키지명 변경 및 AuthController 리팩터링
