+++ src/main/java/roomescape/auth/AuthRequest.java
@@ -1,4 +1,4 @@
-package roomescape.login;
+package roomescape.auth;
 
 public record AuthRequest(
         String email,
