+++ src/main/java/roomescape/auth/AuthResponse.java
@@ -1,4 +1,4 @@
-package roomescape.login;
+package roomescape.auth;
 
 public record AuthResponse(
         String token
