+++ src/main/java/roomescape/auth/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.login.jwt;
+package roomescape.auth.jwt;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
